requirements.txt
rosbags
numpy
ultralytics==8.0.20
Pillow